+++ lib/flop/adapter/ecto/operators.ex
@@ -73,12 +73,14 @@ defmodule Flop.Adapter.Ecto.Operators do
 
   def op_config(:starts_with, false) do
     fragment = like_fragment(quote(do: ^var!(value)))
-    {fragment, prelude(:add_wildcard_suffix), nil}
+    fragment_dynamic = like_fragment_dynamic(quote(do: ^var!(value)))
+    {fragment, fragment_dynamic, prelude(:add_wildcard_suffix), nil}
   end
 
   def op_config(:ends_with, false) do
     fragment = like_fragment(quote(do: ^var!(value)))
-    {fragment, prelude(:add_wildcard_prefix), nil}
+    fragment_dynamic = like_fragment_dynamic(quote(do: ^var!(value)))
+    {fragment, fragment_dynamic, prelude(:add_wildcard_prefix), nil}
   end
 
   def op_config(op, _ilike?), do: op_config(op)
@@ -89,7 +91,12 @@ defmodule Flop.Adapter.Ecto.Operators do
         field(r, ^var!(field)) == ^var!(value)
       end
 
-    {fragment, nil, nil}
+    fragment_dynamic =
+      quote do
+        ^var!(field_dynamic) == ^var!(value)
+      end
+
+    {fragment, fragment_dynamic, nil, nil}
   end
 
   def op_config(:!=) do
@@ -98,7 +105,12 @@ defmodule Flop.Adapter.Ecto.Operators do
         field(r, ^var!(field)) != ^var!(value)
       end
 
-    {fragment, nil, nil}
+    fragment_dynamic =
+      quote do
+        ^var!(field_dynamic) != ^var!(value)
+      end
+
+    {fragment, fragment_dynamic, nil, nil}
   end
 
   def op_config(:>=) do
@@ -107,7 +119,12 @@ defmodule Flop.Adapter.Ecto.Operators do
         field(r, ^var!(field)) >= ^var!(value)
       end
 
-    {fragment, nil, nil}
+    fragment_dynamic =
+      quote do
+        ^var!(field_dynamic) >= ^var!(value)
+      end
+
+    {fragment, fragment_dynamic, nil, nil}
   end
 
   def op_config(:<=) do
@@ -116,7 +133,12 @@ defmodule Flop.Adapter.Ecto.Operators do
         field(r, ^var!(field)) <= ^var!(value)
       end
 
-    {fragment, nil, nil}
+    fragment_dynamic =
+      quote do
+        ^var!(field_dynamic) <= ^var!(value)
+      end
+
+    {fragment, fragment_dynamic, nil, nil}
   end
 
   def op_config(:>) do
@@ -125,7 +147,12 @@ defmodule Flop.Adapter.Ecto.Operators do
         field(r, ^var!(field)) > ^var!(value)
       end
 
-    {fragment, nil, nil}
+    fragment_dynamic =
+      quote do
+        ^var!(field_dynamic) > ^var!(value)
+      end
+
+    {fragment, fragment_dynamic, nil, nil}
   end
 
   def op_config(:<) do
@@ -134,7 +161,12 @@ defmodule Flop.Adapter.Ecto.Operators do
         field(r, ^var!(field)) < ^var!(value)
       end
 
-    {fragment, nil, nil}
+    fragment_dynamic =
+      quote do
+        ^var!(field_dynamic) < ^var!(value)
+      end
+
+    {fragment, fragment_dynamic, nil, nil}
   end
 
   def op_config(:in) do
@@ -143,7 +175,12 @@ defmodule Flop.Adapter.Ecto.Operators do
         field(r, ^var!(field)) in ^var!(value)
       end
 
-    {fragment, nil, nil}
+    fragment_dynamic =
+      quote do
+        ^var!(field_dynamic) in ^var!(value)
+      end
+
+    {fragment, fragment_dynamic, nil, nil}
   end
 
   def op_config(:contains) do
@@ -152,7 +189,12 @@ defmodule Flop.Adapter.Ecto.Operators do
         ^var!(value) in field(r, ^var!(field))
       end
 
-    {fragment, nil, nil}
+    fragment_dynamic =
+      quote do
+        ^var!(value) in ^var!(field_dynamic)
+      end
+
+    {fragment, fragment_dynamic, nil, nil}
   end
 
   def op_config(:not_contains) do
@@ -161,13 +203,19 @@ defmodule Flop.Adapter.Ecto.Operators do
         ^var!(value) not in field(r, ^var!(field))
       end
 
-    {fragment, nil, nil}
+    fragment_dynamic =
+      quote do
+        ^var!(value) not in ^var!(field_dynamic)
+      end
+
+    {fragment, fragment_dynamic, nil, nil}
   end
 
   def op_config(:like) do
     fragment = like_fragment(quote(do: ^var!(value)))
+    fragment_dynamic = like_fragment_dynamic(quote(do: ^var!(value)))
     prelude = prelude(:add_wildcard)
-    {fragment, prelude, nil}
+    {fragment, fragment_dynamic, prelude, nil}
   end
 
   def op_config(:not_like) do
@@ -176,8 +224,13 @@ defmodule Flop.Adapter.Ecto.Operators do
         not unquote(like_fragment(quote(do: ^var!(value))))
       end
 
+    fragment_dynamic =
+      quote do
+        not unquote(like_fragment_dynamic(quote(do: ^var!(value))))
+      end
+
     prelude = prelude(:add_wildcard)
-    {fragment, prelude, nil}
+    {fragment, fragment_dynamic, prelude, nil}
   end
 
   def op_config(:=~) do
@@ -186,8 +239,13 @@ defmodule Flop.Adapter.Ecto.Operators do
         ilike(field(r, ^var!(field)), ^var!(value))
       end
 
+    fragment_dynamic =
+      quote do
+        ilike(^var!(field_dynamic), ^var!(value))
+      end
+
     prelude = prelude(:add_wildcard)
-    {fragment, prelude, nil}
+    {fragment, fragment_dynamic, prelude, nil}
   end
 
   def op_config(:ilike) do
@@ -196,8 +254,13 @@ defmodule Flop.Adapter.Ecto.Operators do
         ilike(field(r, ^var!(field)), ^var!(value))
       end
 
+    fragment_dynamic =
+      quote do
+        ilike(^var!(field_dynamic), ^var!(value))
+      end
+
     prelude = prelude(:add_wildcard)
-    {fragment, prelude, nil}
+    {fragment, fragment_dynamic, prelude, nil}
   end
 
   def op_config(:not_ilike) do
@@ -206,8 +269,13 @@ defmodule Flop.Adapter.Ecto.Operators do
         not ilike(field(r, ^var!(field)), ^var!(value))
       end
 
+    fragment_dynamic =
+      quote do
+        not ilike(^var!(field_dynamic), ^var!(value))
+      end
+
     prelude = prelude(:add_wildcard)
-    {fragment, prelude, nil}
+    {fragment, fragment_dynamic, prelude, nil}
   end
 
   def op_config(:not_in) do
@@ -215,6 +283,12 @@ defmodule Flop.Adapter.Ecto.Operators do
       quote do
         field(r, ^var!(field)) not in ^var!(processed_value) and
           not (^var!(reject_nil?) and is_nil(field(r, ^var!(field))))
+      end
+
+    fragment_dynamic =
+      quote do
+        ^var!(field_dynamic) not in ^var!(processed_value) and
+          not (^var!(reject_nil?) and is_nil(^var!(field_dynamic)))
       end
 
     prelude =
@@ -227,23 +301,25 @@ defmodule Flop.Adapter.Ecto.Operators do
             else: var!(value)
       end
 
-    {fragment, prelude, nil}
+    {fragment, fragment_dynamic, prelude, nil}
   end
 
   def op_config(:like_and) do
     fragment = like_fragment(quote(do: ^substring))
+    fragment_dynamic = like_fragment_dynamic(quote(do: ^substring))
     combinator = :and
     prelude = prelude(:maybe_split_search_text)
 
-    {fragment, prelude, combinator}
+    {fragment, fragment_dynamic, prelude, combinator}
   end
 
   def op_config(:like_or) do
     fragment = like_fragment(quote(do: ^substring))
+    fragment_dynamic = like_fragment_dynamic(quote(do: ^substring))
     combinator = :or
     prelude = prelude(:maybe_split_search_text)
 
-    {fragment, prelude, combinator}
+    {fragment, fragment_dynamic, prelude, combinator}
   end
 
   def op_config(:ilike_and) do
@@ -252,10 +328,15 @@ defmodule Flop.Adapter.Ecto.Operators do
         ilike(field(r, ^var!(field)), ^substring)
       end
 
+    fragment_dynamic =
+      quote do
+        ilike(^var!(field_dynamic), ^substring)
+      end
+
     combinator = :and
     prelude = prelude(:maybe_split_search_text)
 
-    {fragment, prelude, combinator}
+    {fragment, fragment_dynamic, prelude, combinator}
   end
 
   def op_config(:ilike_or) do
@@ -264,10 +345,15 @@ defmodule Flop.Adapter.Ecto.Operators do
         ilike(field(r, ^var!(field)), ^substring)
       end
 
+    fragment_dynamic =
+      quote do
+        ilike(^var!(field_dynamic), ^substring)
+      end
+
     combinator = :or
     prelude = prelude(:maybe_split_search_text)
 
-    {fragment, prelude, combinator}
+    {fragment, fragment_dynamic, prelude, combinator}
   end
 
   def op_config(:starts_with) do
@@ -276,8 +362,13 @@ defmodule Flop.Adapter.Ecto.Operators do
         ilike(field(r, ^var!(field)), ^var!(value))
       end
 
+    fragment_dynamic =
+      quote do
+        ilike(^var!(field_dynamic), ^var!(value))
+      end
+
     prelude = prelude(:add_wildcard_suffix)
-    {fragment, prelude, nil}
+    {fragment, fragment_dynamic, prelude, nil}
   end
 
   def op_config(:ends_with) do
@@ -286,8 +377,13 @@ defmodule Flop.Adapter.Ecto.Operators do
         ilike(field(r, ^var!(field)), ^var!(value))
       end
 
+    fragment_dynamic =
+      quote do
+        ilike(^var!(field_dynamic), ^var!(value))
+      end
+
     prelude = prelude(:add_wildcard_prefix)
-    {fragment, prelude, nil}
+    {fragment, fragment_dynamic, prelude, nil}
   end
 
   # The escape character must be bound rather than written into the fragment
@@ -298,6 +394,17 @@ defmodule Flop.Adapter.Ecto.Operators do
       fragment(
         "? LIKE ? ESCAPE ?",
         field(r, ^var!(field)),
+        unquote(pattern),
+        ^"\\"
+      )
+    end
+  end
+
+  defp like_fragment_dynamic(pattern) do
+    quote do
+      fragment(
+        "? LIKE ? ESCAPE ?",
+        ^var!(field_dynamic),
         unquote(pattern),
         ^"\\"
       )
@@ -329,6 +436,36 @@ defmodule Flop.Adapter.Ecto.Operators do
   defmacro empty(:other) do
     quote do
       is_nil(field(r, ^var!(field)))
+    end
+  end
+
+  defmacro empty_dynamic(:array) do
+    quote do
+      is_nil(^var!(field_dynamic)) or
+        type(^var!(field_dynamic), ^var!(ecto_type)) ==
+          type(^[], ^var!(ecto_type))
+    end
+  end
+
+  # for adapters that store an array as a JSON column
+  defmacro empty_dynamic(:json_array) do
+    quote do
+      is_nil(^var!(field_dynamic)) or
+        fragment("JSON_LENGTH(?) = 0", ^var!(field_dynamic))
+    end
+  end
+
+  defmacro empty_dynamic(:map) do
+    quote do
+      is_nil(^var!(field_dynamic)) or
+        type(^var!(field_dynamic), ^var!(ecto_type)) ==
+          type(^%{}, ^var!(ecto_type))
+    end
+  end
+
+  defmacro empty_dynamic(:other) do
+    quote do
+      is_nil(^var!(field_dynamic))
     end
   end